requirements.txt
anthropic>=0.25.0
gtts>=2.5.1
moviepy>=1.0.3
Pillow>=10.0.0
requests>=2.31.0
APScheduler>=3.10.4
